Rename "torf" module to "builder"
This way the name is library-agnostic, in anticipation of moving from
RoboFab to Defcon.
